pyproject.toml
[project]
name = "rag-task1"
version = "0.1.0"
description = ""
authors = [
    {name = "anjuclaysys",email = "anju.vijayan@claysys.net"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.7.5,<7.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "langchain-qdrant (>=1.1.0,<2.0.0)",
    "qdrant-client (>=1.17.0,<2.0.0)",
    "langchain-huggingface (>=1.2.1,<2.0.0)",
    "llama-index (>=0.14.15,<0.15.0)",
    "llama-index-vector-stores-qdrant (>=0.9.1,<0.10.0)",
    "llama-index-embeddings-huggingface (>=0.6.1,<0.7.0)",
    "google-genai (>=1.66.0,<2.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "fastapi (>=0.135.1,<0.136.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
